+++ packages/charts/src/lib/components/ComposedChartPlaceholder.ts
@@ -1,0 +1,3 @@
+import { ComposedChartPlaceholder } from '../../components/ComposedChart/Placeholder';
+
+export { ComposedChartPlaceholder };